pyproject.toml
[tool.poetry]
name = "cinebot"
version = "0.1.0"
description = ""
authors = ["Alexandra Tabarani <alexandra.tabarani@studenti.luiss.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.13.3"
streamlit = "^1.32.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
